some_new_corrections_in_session_and_with_age
age, first steps in session outlayer deletion, i have an idea, that we can raplace NaN with random numbers, and there are such a function which will give each component exactly that number of random elements, which will not spoil our dataset, just it fill it with not very matched data, but it will be better than most frequent or mean, etc. As it seems to me, what do you think?